Fixed 10th guess bug
Hey!  I noticed your game online and really liked it!  I am a match major so the hint portion really appealed to me and kept it challenging but not impossible.  WHile playing I did notice a bug.  When I got it right on my tenth guess it still said I lost.  This small change fixes it.  Thanks for making a great game!